+++ tests/xml_download/test_utils_cleansing_bulk.py
@@ -1,7 +1,5 @@
-import os
 import pandas as pd
 import pytest
-from os.path import expanduser
 from pathlib import Path
 
 from open_mastr.xml_download.utils_cleansing_bulk import (
@@ -10,22 +8,8 @@ from open_mastr.xml_download.utils_cleansing_bulk import (
     replace_mastr_katalogeintraege,
 )
 
-from tests.conftest import EXISTING_XML_ZIP
 
-# Check if xml file exists
-_xml_file_exists = False
-_xml_folder_path = os.path.join(expanduser("~"), ".open-MaStR", "data", "xml_download")
-if os.path.isdir(_xml_folder_path):
-    for entry in os.scandir(path=_xml_folder_path):
-        if "Gesamtdatenexport" in entry.name and entry.name.endswith(".zip"):
-            _xml_file_exists = True
-
-
-@pytest.mark.skipif(
-    not EXISTING_XML_ZIP,
-    reason="The zipped XML could not be found."
-)
-def test_cleanse_bulk_data(existing_xml_zip_in_output_dir: Path) -> None:
+def test_cleanse_bulk_data(mockup_xml_zip_in_output_dir: Path) -> None:
     df_raw = pd.DataFrame(
         {
             "ID": [0, 1, 2],
@@ -43,37 +27,36 @@ def test_cleanse_bulk_data(existing_xml_zip_in_output_dir: Path) -> None:
 
     pd.testing.assert_frame_equal(
         cleanse_bulk_data(
-            df=df_raw, zipped_xml_file_path=str(existing_xml_zip_in_output_dir), catalog_columns={"Bundesland", "Einheittyp"},
+            df=df_raw,
+            zipped_xml_file_path=str(mockup_xml_zip_in_output_dir),
+            catalog_columns={"Bundesland", "Einheittyp"},
         ),
         df_replaced,
+        check_dtype=False,
     )
 
 
-@pytest.mark.skipif(
-    not EXISTING_XML_ZIP,
-    reason="The zipped XML could not be found."
-)
-def test_replace_mastr_katalogeintraege(existing_xml_zip_in_output_dir: Path) -> None:
+def test_replace_mastr_katalogeintraege(mockup_xml_zip_in_output_dir: Path) -> None:
     df_raw = pd.DataFrame({"ID": [0, 1, 2], "Bundesland": [335, 335, 336]})
     df_replaced = pd.DataFrame(
         {"ID": [0, 1, 2], "Bundesland": ["Bayern", "Bayern", "Bremen"]}
     )
     pd.testing.assert_frame_equal(
         replace_mastr_katalogeintraege(
-            zipped_xml_file_path=str(existing_xml_zip_in_output_dir), df=df_raw, catalog_columns={"Bundesland", "Einheittyp"},
+            zipped_xml_file_path=str(mockup_xml_zip_in_output_dir),
+            df=df_raw,
+            catalog_columns={"Bundesland", "Einheittyp"},
         ),
         df_replaced,
     )
 
 
-@pytest.mark.skipif(
-    not EXISTING_XML_ZIP,
-    reason="The zipped XML could not be found."
-)
-def test_create_katalogwerte_from_bulk_download(existing_xml_zip_in_output_dir: Path) -> None:
+def test_create_katalogwerte_from_bulk_download(
+    mockup_xml_zip_in_output_dir: Path,
+) -> None:
     katalogwerte = create_katalogwerte_from_bulk_download(
-        zipped_xml_file_path=existing_xml_zip_in_output_dir
+        zipped_xml_file_path=mockup_xml_zip_in_output_dir
     )
-    assert type(katalogwerte) == dict
+    assert isinstance(katalogwerte, dict)
     assert len(katalogwerte) > 1000
-    assert type(list(katalogwerte.keys())[0]) == int
+    assert isinstance(list(katalogwerte.keys())[0], int)